refactor(situation): Situation 관련 클래스를 evaluation 패키지로 옮김
